+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 增加异常类型及其相关判断逻辑。
+  - com.dwarfeng.ftp.exception.FtpHandlerStoppedException。
+
 - `FtpConfig` 结构优化。
   - 优化 `FtpConfig` 构造器方法，由 `FtpConfig.Builder` 构造生成时可避免重复的参数校验。
   - 优化 `FtpConfig.Builder` 的参数校验方式，在 build 方法中统一校验参数，避免参数设置顺序改变造成参数校验不通过。